Fixed the relationship between books and users for the ratings, and creating create, delete, and get endpoints.

+++ SuttorLib/Models/BookRating.cs
@@ -13,7 +13,7 @@ namespace SuttorLibrary.Models
 
         public string UserId { get; set; }
 
-        public int Rating { get; set; }
+        public float Rating { get; set; }
         public DateTime CreatedAt { get; set; }
         public string Comment { get; set; } = string.Empty;
     }
